Generalize utility functions and work around circular imports
