+++ imod/mf6/wel.py
@@ -21,11 +21,11 @@ from imod.mf6.interfaces.ipointdatapackage import IPointDataPackage
 from imod.mf6.mf6_wel_adapter import Mf6Wel
 from imod.mf6.package import Package
 from imod.mf6.utilities.dataset import remove_inactive
-from imod.mf6.utilities.grid import create_layered_top
 from imod.mf6.utilities.regrid import RegridderType
 from imod.mf6.validation import validation_pkg_error_message
 from imod.mf6.write_context import WriteContext
 from imod.prepare import assign_wells
+from imod.prepare.layer import create_layered_top
 from imod.schemata import (
     AnyNoDataSchema,
     DTypeSchema,
