+++ imod/prepare/layer.py
@@ -3,6 +3,7 @@ This module contains all kinds of utilities to work with layers.
 """
 
 from imod.typing import GridDataArray
+from imod.typing.grid import zeros_like
 
 
 def get_upper_active_layer_number(active: GridDataArray) -> GridDataArray:
@@ -67,3 +68,14 @@ def get_lower_active_grid_cells(active: GridDataArray) -> GridDataArray:
     layer = active.coords["layer"]
     lower_active_layer = get_lower_active_layer_number(active)
     return layer == lower_active_layer
+
+
+def create_layered_top(bottom: GridDataArray, top: GridDataArray) -> GridDataArray:
+    """
+    Create a top array with layers from a single top array and a full bottom array
+    """
+    new_top = zeros_like(bottom)
+    new_top[0] = top
+    new_top[1:] = bottom[0:-1].values
+
+    return new_top
